+++ pkg/genrec/resources.go
@@ -56,8 +56,11 @@ func (r *Resources) Append(res Resources, err error) error {
 type ResourceOpts struct {
 	IsSensitive    bool
 	DeleteOnChange bool
-	Orphan         bool
-	PatchUpdates   bool
+	// Delete the object if the patch calculation fails.
+	// Useful in scenarios where an object cannot be patched due to the resource containing arbitrary JSON fields.
+	DeleteOnPatchCalculationError bool
+	Orphan                        bool
+	PatchUpdates                  bool
 }
 
 type Resource struct {
@@ -226,14 +229,20 @@ func (rd ResourceDiff) Apply(ctx context.Context, sc k8sutil.SchemedClient, owne
 	log := logr.FromContextOrDiscard(ctx).WithValues("objOp", op, "obj", rd.logInfoMap(sc))
 	switch op {
 	case ResourceDiffOpUpdate:
-		if rd.DeleteOnChange {
-			log.Info("deleting changed object")
-			return client.IgnoreNotFound(sc.Delete(ctx, rd.Observed, client.PropagationPolicy(v1.DeletePropagationBackground)))
-		} else if patchResult, err := om.DefaultPatchMaker.Calculate(rd.Observed, rd.Desired, calcOptions...); err != nil {
+		patchResult, err := om.DefaultPatchMaker.Calculate(rd.Observed, rd.Desired, calcOptions...)
+		if err != nil {
+			// delete objects where there is a diff calculation error
+			if rd.DeleteOnPatchCalculationError {
+				log.Info("deleting object due to patch calculation error")
+				return client.IgnoreNotFound(sc.Delete(ctx, rd.Observed, client.PropagationPolicy(v1.DeletePropagationBackground)))
+			}
 			return err
 		} else if patchResult.IsEmpty() {
 			log.V(2).Info("object is up-to-date")
 			return nil
+		} else if rd.DeleteOnChange {
+			log.Info("deleting object due to delete on change")
+			return client.IgnoreNotFound(sc.Delete(ctx, rd.Observed, client.PropagationPolicy(v1.DeletePropagationBackground)))
 		} else {
 			if rd.IsSensitive {
 				log.Info("updating sensitive object; diff hidden")